requirements.txt
PyMySQL
# PyMySQL==1.0.2
Flask_SQLAlchemy==2.5.1
Flask==1.1.2
gunicorn==20.0.4
Werkzeug==1.0.1
marshmallow==3.13.0
marshmallow-sqlalchemy==0.26.1
flask-marshmallow==0.14.0
PyJWT==2.1.0
pathlib2==2.3.5
boto3==1.18.28
pyshorteners==1.0.1
Flask-Cors==3.0.10